feat: OK! test

+++ includes/config.hpp
@@ -14,7 +14,7 @@ class Config {
         std::string _root;
         std::string _index;
         std::map<int, std::string> _error_page;
-        int _buffer_size;
+        size_t _buffer_size;
         std::vector<std::string> _methods;
         std::map<std::string, Location> _locations;
